Adjust existing commands, pressOffice fields and finish LoadPressMachine. Added description for each command and command to show them

+++ src/main/java/com/nbu/pressofficeapp/models/enums/PaperSize.java
@@ -7,9 +7,9 @@ public enum PaperSize {
     A2(3),
     A1(4);
 
-    public final int sizeMod;
+    public final double sizeMod;
 
-    private PaperSize(int sizeMod){
+    private PaperSize(double sizeMod){
         this.sizeMod = sizeMod;
     }
 }
